% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{Clusters}
\alias{Clusters}
\title{Extract clustering assignments.}
\usage{
Clusters(x, data)
}
\arguments{
\item{x}{An object of class \code{bmix} that represents a fit.}

\item{data}{The data used to compute the fit \code{x}.}
}
\value{
The input \code{data} augmented with a column for the cluster labels and the latent variables.
}
\description{
This function extracts the clustering assignments
and the latent variables from the fit. This function
must receive in input exactly the same data used for the fit.
}
