% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{Parameters}
\alias{Parameters}
\title{Extract the fit parameters of the mixture.}
\usage{
Parameters(x)
}
\arguments{
\item{x}{An object of class \code{bmix} that represents a fit.}
}
\value{
A tibble for the the fit parameters of the mixture.
}
\description{
This function extracts the mixture
parameters as a tibble. The tibble contains
one row per component, a column for the mean of
the mixture component and a column for its overdispersion.
All the Binomial components have a fixed value of 0 for overdispersion..
}
