% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_bmix.R
\name{plot.bmix}
\alias{plot.bmix}
\title{Title}
\usage{
\method{plot}{bmix}(fit, data, coverage = 100, annotation = NULL,
  cex = 1, palette = "Set1", palette.heatmap = "Spectral",
  alpha = 0.3, M.tarone = 1000)
}
\arguments{
\item{alpha}{}
}
\description{
Title
}
