% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot the clustering of the data.}
\usage{
plot_clusters(x, data)
}
\arguments{
\item{x}{An object of class \code{bmix} that represents a fit.}

\item{data}{The data used to compute the fit \code{x}.}
}
\value{
A ggplot object.
}
\description{
This functions plots the data as histogram, coloured
according to the hard clustering assignments. This function
must receive in input exactly the same data used for the fit.
}
