% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_bmix.R
\name{plot_density}
\alias{plot_density}
\title{Title}
\usage{
plot_density(fit, data, coverage = 100, annotation = NULL, cex = 1,
  palette = "Set1", palette.heatmap = "Spectral", alpha = 0.3,
  M.tarone = 1000)
}
\arguments{
\item{M.tarone}{}
}
\description{
Title
}
