% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_selection.R
\name{plot_model_selection}
\alias{plot_model_selection}
\title{Plot the grid of model selection for the mixture.}
\usage{
plot_model_selection(x)
}
\arguments{
\item{x}{An object of class \code{bmix} that represents a fit.}
}
\value{
A ggplot object.
}
\description{
This functions plots the score of each combination of parameters
for this mixture.
}
