% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_bmix.R
\name{print.bmix}
\alias{print.bmix}
\title{Print a \code{bmix} object.}
\usage{
\method{print}{bmix}(x, ...)
}
\arguments{
\item{x}{A \code{bmix} object.}

\item{...}{S3 parameters.}
}
\value{
Nothing, just print to screen some information about the mixture.
}
\description{
Print to screen some information about the mixture.
}
