% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_string.R
\name{to_string}
\alias{to_string}
\title{Export a tibble with the parameters fits for this model}
\usage{
to_string(x)
}
\arguments{
\item{x}{An object of class \code{bmix} that represents a fit.}
}
\value{
A `tibble` for the statistcis.
}
\description{
This functions extracts a tibble reporting the parameters fits for
the input model (clusters, means of clusters, dispersion, likelihood
score, cluster sizes etc.).
}
\examples{
# The same dataset used in the package vignette
data = data.frame(successes = c(rbinom(30, 100, .4), rbinom(70, 100, .7)), trials = 100)

# BMix fit with default parameters
x = bmixfit(data)

to_string(x)
}
