% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_bmix.R
\name{plot.bmix}
\alias{plot.bmix}
\title{Plot a \code{bmix} object.}
\usage{
plot.bmix(x, data)
}
\arguments{
\item{x}{An object of class \code{bmix} that represents a fit.}

\item{data}{The data used to compute the fit \code{x}.}
}
\value{
A \code{cowplot} figure.
}
\description{
Assemble a multi-panel plot using the plotting functions
of the \code{Bmix} package, with default parameters.
}
