% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_bmix.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Title}
\usage{
plot_clusters(fit, data, palette = "Set1",
  palette.heatmap = "Spectral", alpha = 0.3, M.tarone = 1000)
}
\arguments{
\item{M.tarone}{}
}
\value{

}
\description{
Title
}
