% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot the density of the mixture.}
\usage{
plot_density(x, data, trials = round(median(data[, 2])))
}
\arguments{
\item{x}{An object of class \code{bmix} that represents a fit.}

\item{data}{The data used to compute the fit \code{x}.}

\item{trials}{To compute the density a number of trials must be
fixed. This parameter represents exactly that. By default the median
number of trials in the input \code{data} are taken.}
}
\value{
A ggplot object.
}
\description{
This functions plots the density of the fit, coloured
according to the mixture components. This function
must receive in input exactly the same data used for the fit.
}
