% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_selection.R
\name{plot_model_selection}
\alias{plot_model_selection}
\title{Plot the grid of model selection for the mixture.}
\usage{
plot_model_selection(x)
}
\arguments{
\item{x}{An object of class \code{bmix} that represents a fit.}
}
\value{
A ggplot object.
}
\description{
This functions plots the score of each combination of parameters
for this mixture.
}
\examples{
# The same dataset used in the package vignette
data = data.frame(successes = c(rbinom(30, 100, .4), rbinom(70, 100, .7)), trials = 100)

# BMix fit with default parameters
x = bmixfit(data)

plot_model_selection(x)
}
