% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_bmix.R
\name{print.bmix}
\alias{print.bmix}
\title{Print a \code{bmix} object.}
\usage{
\method{print}{bmix}(x, ...)
}
\arguments{
\item{x}{A \code{bmix} object.}

\item{...}{S3 parameters.}
}
\value{
Nothing, just print to screen some information about the mixture.
}
\description{
Print to screen some information about the mixture.
}
\examples{
# The same dataset used in the package vignette
data = data.frame(successes = c(rbinom(30, 100, .4), rbinom(70, 100, .7)), trials = 100)

# BMix fit with default parameters
x = bmixfit(data)

print(x)
}
